[REFACTOR] BookDetailResponse DTO 가격 타입 변경 (#247)
* feat: BookDetailResponse 가격 타입 변경
- int -> BigDecimal

* feat: BookDetailResponse 가격 타입 변경
- 테스트 코드 생성자 매개변수 수정

* feat: BookDetailResponse 가격 타입 변경
- MapStruct int->BigDecimal

---------

+++ src/main/java/com/nhnacademy/illuwa/d_book/book/dto/response/BookDetailResponse.java
@@ -2,6 +2,7 @@ package com.nhnacademy.illuwa.d_book.book.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -17,8 +18,8 @@ public class BookDetailResponse {
     private String publisher;
     private LocalDate publishedDate;
     private String isbn;
-    private int regularPrice;
-    private int salePrice;
+    private BigDecimal regularPrice;
+    private BigDecimal salePrice;
     private boolean isGiftWrap;
     private String imgUrl;
 }
